Get build process working correctly
