+++ projects/packages/forms/src/modules/file-field-interactive/block.json
@@ -1,11 +1,11 @@
 {
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 3,
-	"name": "jetpack/form-field-interactive",
+	"name": "jetpack/file-field-interactive",
 	"supports": {
 		"interactivity": true
 	},
 	"textdomain": "jetpack-forms",
 	"title": "File Field",
-	"viewScriptModule": "file:./view.js"
+	"viewScriptModule": "file:./index.js"
 }
